// src/app/modules/score-submission/pages/update-kpi-progress/enums.ts
export enum SubmissionState {
  NotDefined = 0,
  PendingForSubmission = 1,
  Missed = 2,
  Achieved = 3,
  OverAchieved = 4,
  NotDue = 5,
}

export enum TrackStatus {
  OffTrack = 0,
  OnTrack = 1
}

export enum FrequenclyMode {
  Monthly = 1,
  Quarterly = 2,
  Biyearly = 3,
  Yearly = 4
}

export enum MeasurementMethodStatus {
  Percentage = 1,
  Number = 2,
  YesOrNo = 3,
  Currency = 4,
  DateKpi = 5,
}